fixed SDF16 not rendering (2 bugs)

+++ Internal/Harfbuzz/Bitmap/SdfRasterizer.cs
@@ -35,7 +35,7 @@ namespace TextMeshDOTS.HarfBuzz.Bitmap
                 var distSq         = distSqs[i];
                 var signedDistance = sign * math.min(math.sqrt(distSq) / spread, 1f);  // in [-1, 1] range
                 var scaled         = (signedDistance + 1f) * 65535f / 2f;  // unorm correction
-                buffer[i]          = (byte)scaled;
+                buffer[i]          = (ushort)scaled;
             }
             PaintUtils.rasterizeSDFMarker.End();
         }
